+++ src/components/Form/DatePicker.field.tsx
@@ -1,0 +1,31 @@
+import type { IFieldProps } from './types'
+import { StyledDatePicker, ErrorMessage } from './elements'
+
+interface IDatePickerFieldProps extends IFieldProps {
+  customChange?: (location) => void
+}
+
+export const DatePickerField = ({
+  input,
+  meta,
+  customChange,
+  ...rest
+}: IDatePickerFieldProps) => (
+  <>
+    <StyledDatePicker
+      invalid={meta.error && meta.touched}
+      {...input}
+      {...rest}
+      onChange={(date) => {
+        input.onChange(date)
+        if (customChange) {
+          customChange(date)
+        }
+        input.onBlur()
+      }}
+    />
+    {meta.error && meta.touched ? (
+      <ErrorMessage>{meta.error}</ErrorMessage>
+    ) : null}
+  </>
+)
