+++ src/components/Form/LocationSearch.field.tsx
@@ -1,4 +1,4 @@
-import type { IFieldProps } from './Fields'
+import type { IFieldProps } from './types'
 import OsmGeocoding from '../OsmGeocoding/OsmGeocoding'
 import { logger } from 'src/logger'
 import { getLocationData } from 'src/utils/getLocationData'
