+++ src/pages/Events/Content/EventsCreate/EventsCreate.tsx
@@ -4,7 +4,8 @@ import { Form, Field } from 'react-final-form'
 import arrayMutators from 'final-form-arrays'
 import TEMPLATE from './Template'
 import type { UploadedFile } from 'src/pages/common/UploadedFile/UploadedFile'
-import { InputField, DatePickerField } from 'src/components/Form/Fields'
+import { InputField } from 'src/components/Form/Input.field'
+import { DatePickerField } from 'src/components/Form/DatePicker.field'
 import { Button } from 'oa-components'
 import type { EventStore } from 'src/stores/Events/events.store'
 import Heading from 'src/components/Heading'
