+++ src/pages/Howto/Content/Common/Howto.form.tsx
@@ -8,7 +8,8 @@ import createDecorator from 'final-form-calculate'
 import type { IHowtoFormInput } from 'src/models/howto.models'
 import Text from 'src/components/Text'
 import type { UploadedFile } from 'src/pages/common/UploadedFile/UploadedFile'
-import { InputField, TextAreaField } from 'src/components/Form/Fields'
+import { InputField } from 'src/components/Form/Input.field'
+import { TextAreaField } from 'src/components/Form/TextArea.field'
 import { SelectField } from 'src/components/Form/Select.field'
 import { HowtoStep } from './HowtoStep.form'
 import { Button } from 'oa-components'
