+++ src/pages/Settings/content/formSections/ChangePassword.form.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Field, Form } from 'react-final-form'
 import { Button } from 'oa-components'
-import { InputField } from 'src/components/Form/Fields'
+import { InputField } from 'src/components/Form/Input.field'
 import type { UserStore } from 'src/stores/User/user.store'
 import Text from 'src/components/Text'
 import theme from 'src/themes/styled.theme'
