+++ src/pages/Settings/content/formSections/Fields/Link.field.tsx
@@ -1,7 +1,7 @@
 import { Component } from 'react'
 import { COM_TYPE_MOCKS } from 'src/mocks/Selectors'
 import { Field } from 'react-final-form'
-import { InputField } from 'src/components/Form/Fields'
+import { InputField } from 'src/components/Form/Input.field'
 import { Button } from 'oa-components'
 import { Modal } from 'src/components/Modal/Modal'
 import Text from 'src/components/Text'
