+++ src/pages/Settings/content/formSections/MemberMapPin.section.tsx
@@ -4,7 +4,7 @@ import Heading from 'src/components/Heading'
 import { Field } from 'react-final-form'
 import Text from 'src/components/Text'
 import { Button } from 'oa-components'
-import { TextAreaField } from 'src/components/Form/Fields'
+import { TextAreaField } from 'src/components/Form/TextArea.field'
 import { Box, Flex } from 'theme-ui'
 import { FlexSectionContainer, ArrowIsSectionOpen } from './elements'
 import { MAP_GROUPINGS } from 'src/stores/Maps/maps.groupings'
