+++ src/pages/Settings/content/formSections/UserInfos.section.tsx
@@ -3,7 +3,8 @@ import { Field } from 'react-final-form'
 import Heading from 'src/components/Heading'
 import Text from 'src/components/Text'
 import Flex from 'src/components/Flex'
-import { InputField, TextAreaField } from 'src/components/Form/Fields'
+import { InputField } from 'src/components/Form/Input.field'
+import { TextAreaField } from 'src/components/Form/TextArea.field'
 import { Button } from 'oa-components'
 import theme from 'src/themes/styled.theme'
 import { FieldArray } from 'react-final-form-arrays'
