+++ src/pages/Settings/content/formSections/WorkspaceMapPin.section.tsx
@@ -3,7 +3,7 @@ import { observer, inject } from 'mobx-react'
 import Heading from 'src/components/Heading'
 import { Field } from 'react-final-form'
 import Text from 'src/components/Text'
-import { TextAreaField } from 'src/components/Form/Fields'
+import { TextAreaField } from 'src/components/Form/TextArea.field'
 import { Box, Flex, Link } from 'theme-ui'
 import { FlexSectionContainer, ArrowIsSectionOpen } from './elements'
 import { MAP_GROUPINGS } from 'src/stores/Maps/maps.groupings'
