+++ src/pages/SignIn/SignIn.tsx
@@ -8,7 +8,7 @@ import Heading from 'src/components/Heading'
 import { Button } from 'oa-components'
 import Text from 'src/components/Text'
 import { Link } from 'src/components/Links'
-import { InputField } from 'src/components/Form/Fields'
+import { InputField } from 'src/components/Form/Input.field'
 import { inject, observer } from 'mobx-react'
 import { Form, Field } from 'react-final-form'
 import type { UserStore } from 'src/stores/User/user.store'
