+++ src/pages/SignUp/SignUp.tsx
@@ -8,7 +8,7 @@ import { Button } from 'oa-components'
 import Text from 'src/components/Text'
 import { Link } from 'src/components/Links'
 import { Form, Field } from 'react-final-form'
-import { InputField } from 'src/components/Form/Fields'
+import { InputField } from 'src/components/Form/Input.field'
 import { inject, observer } from 'mobx-react'
 import type { UserStore } from 'src/stores/User/user.store'
 import type { RouteComponentProps } from 'react-router'
